version 0.4.3

+++ app/Http/Controllers/AdminController.php
@@ -10,7 +10,10 @@ use Lang;
 use aria2;
 use App\User;
 use Hash;
+use Validator;
 use main;
+use Illuminate\Support\Facades\Config;
+use Input;
 
 class AdminController extends Controller
 {
@@ -24,13 +27,131 @@ class AdminController extends Controller
         //
     }
 
-    public function stat(){
+    public function stat()
+    {
         $main = new main();
         if (!$main->aria2_online()) return view('errors.general', array('error_title' => 'ERROR 10002', 'error_message' => 'Aria2c is not running!'));
 
-        return view('tools.stats', array('main' => $main));
-
+        $aria2 = new aria2();
+        return view('tools.stats', array('main' => $main, 'aria2' => $aria2));
     }
+
+
+
+
+    public function post_stat(Request $request)
+    {
+        if ($request->ajax()) {
+            if (isset($request['gs'])) {
+                $main = new main();
+                if (!$main->aria2_online())
+                    return response()->json(['ERROR 10002' => 'Aria2 is not running!']);
+
+                $aria2 = new aria2();
+
+                return response()->json([
+                    'speed' => $main->formatBytes($aria2->getGlobalStat()['result']['downloadSpeed'], 3),
+                    'speed_b' => round($aria2->getGlobalStat()['result']['downloadSpeed'] / 1024, 0),
+                    'time' => time(),
+                    'numActive' => $aria2->getGlobalStat()['result']['numActive'],
+                    'numStopped' => $aria2->getGlobalStat()['result']['numStopped'],
+                    'numWaiting' => $aria2->getGlobalStat()['result']['numWaiting']
+                ]);
+            }elseif(isset($request['lf'])){
+                $table = [];
+                $files = DB::table('download_list')
+                    ->leftJoin('users', 'users.id', '=', 'download_list.user_id')
+                    ->select('download_list.*', 'users.username')
+                    ->orderBy('id','DEC')
+                    ->take(20)
+                    ->get();
+
+                $main = new main();
+                foreach($files as $file) {
+                    if ($file->state == null)
+                        $status = Lang::get('messages.in_queue');//in_queue
+                    elseif ($file->state == 0)
+                        $status = Lang::get('messages.finished');
+                    elseif ($file->state == -1)
+                        $status = Lang::get('messages.downloading');
+                    elseif ($file->state == -2)
+                        $status = Lang::get('messages.paused');
+                    elseif ($file->state == -3)
+                        $status = Lang::get('messages.deleted');
+                    else
+                        $status = Lang::get('messages.error_id', ['id' => $file->state]);
+
+                    $table[] = [
+                        'id' => $file->id,
+                        'file_name' => $file->file_name,
+                        'length' => $main->formatBytes($file->length, 1),
+                        'date_added' => $file->date_added,
+                        'details' => url('/files/' . $file->id),
+                        'details_t' => Lang::get('messages.details'),
+                        'state' => $status,
+                        'username' => $file->username,
+                        'username_l' => url('tools/users/' . $file->username)
+                    ];
+                }
+                return response()->json($table);
+            }
+        }
+
+        return redirect::back();
+    }
+
+
+
+
+    public function user_details_credits($user_name)
+    {
+        $user = User::where('username', '=', $user_name)->first();
+
+        $tracks = DB::table('credit_log')
+            ->select('credit_log.*', 'users.username')
+            ->join('users', 'credit_log.agent', '=', 'users.id')
+            ->where('user_id', '=', $user->id)
+            ->get();
+
+        $main = new main();
+        return view('tools.user_credits', array('main' => $main, 'user' => $user, 'tracks' => $tracks));
+    }
+
+
+
+
+    public function postuser_details_credits(Request $request, $user_name)
+    {
+        $input = $request->only('new_credit');
+
+        $this->validate($request, [
+            'new_credit' => 'required|numeric'
+        ]);
+        if ($input['new_credit'] < 0) {
+            return redirect::back()->withErrors(Lang::get('errors.neg_num'));
+        }
+
+        $input['new_credit'] *=  1024 * 1024 * 1024;
+        $user = User::where('username', '=', $user_name)->first();
+        $old_credit = $user->credit;
+
+        DB::table('users')
+            ->where('username', $user_name)
+            ->update(['credit' => $input['new_credit']]);
+
+        DB::table('credit_log')->insert(
+            array(
+                'user_id' => $user->id,
+                'credit_change' =>  $input['new_credit'] - $old_credit,
+                'agent' => Auth::user()->id,
+            )
+        );
+
+        return redirect::back();
+    }
+
+
+
 
     public function aria2console()
     {
@@ -44,11 +165,13 @@ class AdminController extends Controller
     }
 
 
+
+
     public function post_aria2console(Request $request)
     {
         $input = $request->only('function', 'param');
 
-        if (empty($_SERVER['HTTP_X_REQUESTED_WITH']) && strtolower($_SERVER['HTTP_X_REQUESTED_WITH']) != 'xmlhttprequest') {
+        if (!$request->ajax()) {
             $this->validate($request, [
                 'function' => 'required'
             ]);
@@ -57,7 +180,7 @@ class AdminController extends Controller
         $functions = array('addUri', 'addTorrent', 'addMetalink', 'remove', 'forceRemove', 'pause', 'pauseAll', 'forcePause', 'forcePauseAll', 'unpause', 'unpauseAll', 'tellStatus', 'getUris', 'getFiles', 'getPeers', 'getServers', 'tellActive', 'tellWaiting', 'tellStopped', 'changePosition', 'changeUri', 'getOption', 'changeOption', 'getGlobalOption', 'changeGlobalOption', 'getGlobalStat', 'purgeDownloadResult', 'removeDownloadResult', 'getVersion', 'getSessionInfo', 'shutdown', 'forceShutdown', 'saveSession', 'multicall');
 
         if (!in_array($input['function'], $functions)) {
-            if (!empty($_SERVER['HTTP_X_REQUESTED_WITH']) && strtolower($_SERVER['HTTP_X_REQUESTED_WITH']) == 'xmlhttprequest') {
+            if ($request->ajax()) {
                 return response()->json(['Error' => 'The function does not exist in Aria2 functions.']);
             } else {
                 return redirect::back()->withErrors('The function does not exist in Aria2 functions.');
@@ -66,7 +189,7 @@ class AdminController extends Controller
 
         $main = new main();
         if (!$main->aria2_online())
-            if (!empty($_SERVER['HTTP_X_REQUESTED_WITH']) && strtolower($_SERVER['HTTP_X_REQUESTED_WITH']) == 'xmlhttprequest') {
+            if ($request->ajax()) {
                 return response()->json(['ERROR 10002' => 'Aria2c is not running!']);
             } else {
                 return view('errors.general', array('error_title' => 'ERROR 10002', 'error_message' => 'Aria2c is not running!'));
@@ -81,7 +204,7 @@ class AdminController extends Controller
 
         $res = call_user_func_array(array($aria2, 'JSON_INPUT' . $input['function']), array($params));
 
-        if (!empty($_SERVER['HTTP_X_REQUESTED_WITH']) && strtolower($_SERVER['HTTP_X_REQUESTED_WITH']) == 'xmlhttprequest') {
+        if ($request->ajax()) {
             return response()->json($res);
         }
 
@@ -90,9 +213,13 @@ class AdminController extends Controller
             ->with('result', $res);
     }
 
-    public function postuser_details($username){
+
+
+
+    public function postuser_details($username)
+    {
         if (!empty($username) || !isset($_POST['action'])) {
-            if ($_POST['action'] == 'delete') {
+            if ($_POST['action'] == 'delete' && Config::get('leech.user_delete') == true) {
                 $atu = Auth::user()->username;
                 $user = User::where('username', '=', $username);
                 $user->delete();
@@ -119,6 +246,10 @@ class AdminController extends Controller
 
     }
 
+
+
+
+
     public function user_details($user_name)
     {
 
@@ -135,9 +266,10 @@ class AdminController extends Controller
             SUM(length) as length_sum,
             COUNT(*) as total_files_deleted,
             (select COUNT(*) from download_list where deleted=0 AND user_id = ' . $users->id . ') as total_files,
+            (select SUM(`length`) from download_list where state <> 0 AND deleted = 0 AND user_id = ' . $users->id . ') as queue_credit,
             (select COUNT(*) from download_list where state > 0 AND deleted=0 AND user_id = ' . $users->id . ') as total_error_files,
             (select COUNT(*) from download_list where state > 0 AND user_id = ' . $users->id . ') as total_error_files_deleted,
-            (select COUNT(*) from download_list where state is NULL AND user_id = ' . $users->id . ') as total_download_queue
+            (select COUNT(*) from download_list where state <> 0 AND deleted = 0 AND user_id = ' . $users->id . ') as total_download_queue
 
             '))
             ->where('user_id', '=', $users->id)
@@ -150,15 +282,116 @@ class AdminController extends Controller
         return view('tools.userdetails', array('user_files' => $user_files, 'user' => $users, 'userd' => $user_detailed, 'main' => $main));
     }
 
+
+
+
     public function users()
     {
         $main = new main();
 
+        $page = Input::get('page');
+        if (!$page)
+            $page = 1;
+
+        $users_count = DB::table('users')->count();
+        $skip = ($page - 1) * 20;
+        $take = 20;
+
+        if ($page == 'all')
+        {
+            $skip = 0;
+            $take = $users_count;
+        }
         $users = DB::table('users')
+            ->where('id', '>' , 0)
+            ->skip($skip)
+            ->take($take)
             ->get();
 
-        return view('tools.users', array('users' => $users, 'main' => $main));
+        return view('tools.users', array('users' => $users, 'main' => $main, 'users_count' => $users_count));
     }
 
+
+
+    public function post_users(Request $request)
+    {
+        $main = new main();
+
+        $input = $request->only('id', 'first_name', 'last_name', 'username', 'email', 'public', 'torrent', 'active');
+        $post_back = $input;
+
+        if ($input['id'] == trim('')) $input['id'] = '%';
+        if ($input['first_name'] == trim('')) $input['first_name'] = '%';
+        if ($input['last_name'] == trim('')) $input['last_name'] = '%';
+        if ($input['username'] == trim('')) $input['username'] = '%';
+        if ($input['email'] == trim('')) $input['email'] = '%';
+
+        if ($input['torrent'] == 2) $input['torrent'] = 1;
+        elseif ($input['torrent'] == 3) $input['torrent'] = 0;
+        else $input['torrent'] = '%';
+
+        if ($input['public'] == 2) $input['public'] = 1;
+        elseif ($input['public'] == 3) $input['public'] = 0;
+        else $input['public'] = '%';
+
+        if ($input['active'] == 2) $input['active'] = 1;
+        elseif ($input['active'] == 3) $input['active'] = 0;
+        else $input['active'] = '%';
+
+//        return $input['active'];
+
+//        return $input['id'];
+
+        $users = DB::table('users')
+            ->where('id', '>' , 0)
+            ->where('id', 'like' , '%' . $input['id'] . '%')
+            ->where('first_name', 'like' , '%' . $input['first_name'] . '%')
+            ->where('last_name', 'like' , '%' . $input['last_name'] . '%')
+            ->where('username', 'like' , '%' . $input['username'] . '%')
+            ->where('email', 'like' , '%' . $input['email'] . '%')
+            ->where('active', 'like' , '%' . $input['active'] . '%')
+            ->where('public', 'like' , '%' . $input['public'] . '%')
+            ->where('torrent', 'like' , '%' . $input['torrent'] . '%')
+            ->get();
+//        'id' => ,
+
+        return view('tools.users', array(
+            'users' => $users,
+            'main' => $main,
+            'page' => false,
+            'id' => $post_back['id'],
+            'username' => $post_back['username'],
+            'last_name' => $post_back['last_name'],
+            'first_name' => $post_back['first_name'],
+            'email' => $post_back['email'],
+            'active' => $post_back['active'],
+            'public' => $post_back['public'],
+            'torrent' => $post_back['torrent'],
+        ));
+    }
+
+
+
+    public function files(Request $request)
+    {
+        $main = new main();
+
+        $page = Input::get('page');
+        if (!$page)
+            $page = 1;
+
+        $files_count = DB::table('download_list')->count();
+
+        $files = DB::table('download_list')
+            ->leftJoin('users', 'users.id', '=', 'download_list.user_id')
+            ->select('download_list.*', 'users.username')
+            ->where('download_list.id', '>' , 0)
+            ->orderBy('id','DEC')
+            ->skip(($page - 1) * 20)
+            ->take(20)
+            ->get();
+
+        return view('tools.all_files', ['files' => $files, 'main' => $main, 'files_count' => $files_count]);
+    }
 
 }
